[Feature #161378879] move midlleware
