+++ app/api/v2/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+apiv1 = Blueprint('apiv1', __name__)
+
+api = Api(apiv1, prefix='/api/v2')
+
